requirements.txt
aiogram==3.4.1
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
redis==5.0.7
